Add scripts for virtual environments
This commit adds two rudimentary scripts for interacting with
Conda environments and Python virtual environments created with
`venv`. Both scripts follow the same pattern: a table of new
environment variables is created via the script and the user
must activate it by loading it with `load-env`:

`load-env (conda-env name-of-env)`

The environments are deactivated by sourcing a script, which
uses `unlet-env` to remove the environment variables, and restores
the old path.

In both cases, nesting virtual environments is not supported. As
soon as an environment is activated, all information from the previous
one is overwritten.