add desired values to the template
This commit adds the background, the foreground and the cursor
colors to the themes as commented lines.
These do not fit into any nushell theme as they should be set by the
terminal emulator, but they are included here for completeness.

Addresses https://github.com/nushell/nu_scripts/pull/286#issuecomment-1230814914