Convert panache-git from script to module
In Nushell 0.68, the `source` command has been deprecated.
Going forward, the way to utilize custom commands is
to put them into a module and import them with `use`.

This means that `panache-git.nu` had to be converted
from a script to a module so that the main prompt command
could be imported by a user's Nushell environment config.