Remove the failing CI on the main branch
As we don't have a path forward to make sure all files are fixed and
will be maintained. (and the file detection itself is reliable)
Disable the `main` branch (and nightly run) for now.

This will keep the CI for PRs so at least added scripts pass the current
nu version