panache-git: do not print stderr from Git commands
Nushell 0.92 changed how external commands operate in pipelines:
https://www.nushell.sh/blog/2024-04-02-nushell_0_92_0.html

After updating to Nushell 0.92, panache-git would print errors
when the current working directory was not a Git repository:
"fatal: not a git repository (or any of the parent directories): .git"

This change properly handles stderr from Git commands
according to Nushell's updated external command behavior.

Also, Nushell 0.91 introduced "is-not-empty", so places in the code
where a string was checked for being not-empty or not-contains
were simplified a bit.

Finally, the main command in the module was renamed to "main"
to make it simpler to import and use in config.