Added Git completions for the grep subcommand
Notes for further improvement

The following flags/parameters have been omitted for now:
- --open-files-in-pager(-O)
- --color
- <trees>...

`--open-files-in-pager(-O)` work as intended when not passing parameters
to use a non-default pager. When passing a non-default pager, the
arguments need to be passed as a literal string to Git as either
`--open-files-in-pager=<pager>` or `-O<pager>`. I need some help with
how to handle that, although it's likely a rare use-case.

`--color` is similar, in that Git also expects a literal string with
`--color=<option>`.

`<trees>` have been omitted for now, due to that being an advanced way
to use `git grep`, and I'm not sure to how to provide good completion
options for that at this time.