Remove multi-line flag from regexes (#903)
Minor fix on `str dedent` - I erroneously used the regex multi-line flag
`(?m)` in several patterns where it shouldn't have been. For the most
part, this was harmless, but it did surface an error when I was working
on updating the themes.

Fixed these and added two new tests that would have caught the issue.