update oh-my.nu to respect use_ansi_coloring config setting (#1003)
A quick update to allow the oh-my.nu prompt to respect the
$env.config.use_ansi_coloring by way of a new command called `config
use-colors`. This way the prompt will be rendered correctly as far as
ansi escapes.

This file was also formatted with `topiary` which makes up for the rest
of the changes.

This will only work on nushell version 0.101.1 and above.