Rename path replace-extension to path with-extension, add with-stem and with-parent (#1011)
Extends #1002. Renames `path replace-extension` to `path
with-extension`, in following with other languages
([Rust](https://doc.rust-lang.org/std/path/struct.Path.html#method.with_extension),
[Python](https://docs.python.org/3/library/pathlib.html#pathlib.PurePath.with_suffix)),
and adds `path with-stem` and `path with-parent`. Also moves the `path`
module into `std-rfc` so it can be used like `use std-rfc/path`.

Adds a private helper function, `with-field`, that `with-extension`,
`with-stem`, and `with-parent` can use. These can each be dead simple
functions, while giving users more options for path manipulation.

The motivation for separate `with-extension`, `with-stem`, and
`with-parent` functions, rather than a more general function like `path
with` is the following:
- `with-extension` has special behavior for stripping periods
- you can tab-complete `path with<TAB>` to immediately see all the
possible options
- you can't accidentally pass an invalid field to `path with` 
- there can be separate examples for `with-extension`, `with-stem`,
`with-parent` for only the relevant functionality