panache-git: exclude prompt indicator so user can choose their own (#1012)
Previously, panache-git included a prompt indicator (`> `) out of the
box, and the advice for the user was to remove Nushell's own
configurable prompt indicator by setting it to the empty string.

However, this conflicted with some of Nushell's intended functionality,
such as the ability to use a different prompt indicator in each vi mode.

With this change, panache-git does not include a prompt indicator. The
advice is for the user to use the prompt indicator they like, but to
reset the color so it better matches panache-git's colors.

This is something of a breaking change, in that users who update to the
latest panache-git.nu without making corresponding changes to their
Nushell config might see two prompt indicators (`> > `), but my hope is
that giving users (especially vi mode users) more flexibility with their
prompt indicators will be worth any temporary confusion.