key-value module for std-rfc (#965)
`std-rfc/kv` is a straightforward but flexible interface for setting and
getting key-value pairs.

## Use-cases

* Ergonomically assign the result of a pipeline to a "variable". Just
"Up Arrow" and append `| kv set foo`.
* Use mid-pipeline to assign a "variable" and yet still continue the
pipeline
* Use mid-pipeline to inspect the state (like the `inspect` command) and
examine the results via `kv list` or `kv get` after the command
completes.
* Chaining assignments/setters
* Set universal variables once and access them even after the shell
exits (or in other simultaneously running shells).

## Features

* Values can be any Nushell type other than a closure. Values are
converted to and from nuons that are stored in a SQLite database.
* The module's commands can operate on either an in-memory database
(using `stor`) or on-disk (`into sqlite`).
* Includes a hook that enables "universal variables" similar to that of
the Fish shell. Universal variables are environment variables that are
immediately updated and available in *all* Nushell sessions that are
running the hook. Since they are stored in an on-disk SQLite database,
they also persist when the shell exits.
* Because kv pairs are stored as rows in a database, they can be
removed, unlike normal variables.
* kv pairs are easily converted to a record using `| transpose -dr`. The
resulting record is, of course, easily converted to environment
variables using `load-env`.
* Assignment can come from either pipeline input or a positional
parameter. When both are provided, the positional parameter is preferred
so that `$in` can be used.
* A closure can be used to modify the pipeline input before storing.
* Can optionally return either the pipeline input (default), the value
that was set, or the entire store back to continue the pipeline.