Add AeroSpace completions (#1080)
First attempt at a full custom completion! Got all the base commands set
and documented. Most arguments are fleshed out, but I haven't figured
out a good solution for subcommands that have exclusive arguments, or
how to progressively show subcommands with their own set of externs.

I've put the completed-but-not-implemented work in a separate file, to
be incorporated later as I figure out the process to do so.

---------

Co-authored-by: Joshua Nussbaum <67082011+joshuanussbaum@users.noreply.github.com>