Fix $env.PATH for direnv (#1083)
While using direnv with nushell, I noticed that my autocompletion for
external programs stopped working as soon as I entered a directory and
direnv loaded.

Turns out, direnv exports $PATH as a string, and nushell stores $PATH as
a table and expects it to be one. Worse, nu just silently fails any
operations with the $PATH (except for running programs) instead of
reporting an error.

This PR fixes that by adding a line in the direnv config.nu script that
parses that string into the appropriate table.
Testing it on my system (NixOS, nushell, direnv + lorri) it seems to
work perfectly fine and the $PATH works again.