Port before_v0.60 parse_aws_s3_ls.nu
Tested with https://docs.aws.amazon.com/cli/latest/reference/s3/ls.html example output

Note that the script has a placeholder parameter `s3://your-bucket-and-path`.
The script is moved into the `sourced` folder like all the other ported scripts from the cool_oneliners folder before it,
but sourcing it as-is is not meaningful.

This ports the last file in `before_v0.60`, consequently resolving #221.