add `capture-foreign-env` module (#1099)
This is a modified version of [capture-foreign-env command from nushell
wiki](https://www.nushell.sh/cookbook/foreign_shell_scripts.html#detailed-explanation-of-capture-foreign-env).
Unlike the wiki version It relies on `null_byte` instead of `newline` as
a delimiter for environment variables dumped by the `env` command which
allows it to support multi-line foreign environment variables.