feat(gh): list my stars (#1109)
Sometimes you know you starred something long time ago, the you go to
github.com your stars and search on the bar.

That's all good but wouldn't it be great to do that from the comfort of
nushell?

This pr does that.

You can test it with
```nu
gh my stars | explore
```