feat: ani-cli completions rework (#1149)
most of the changes are in the commit descriptions, but briefly it just
a rework of ani-cli completions. there were some outdated (or just
non-existent from the start) parameters, some errors and i didn't like
the styling. while at it, i added more functionality to the completions,
examples and remaining options.