Disable recipe argument completion
This completion doesn't do much, just shows the names of the recipe
arguments. However, it prevents autocompletion for all the other
parameters of `just`.

I wonder if there could be a nushell bug here. When we have rest
parameters with autocompletion function, the completion for other
parameters does not work any more (nushell 0.106.1). I tried, it
is still the case even if the completion function returns null.

Nevertheless, at the moment it is better to get rid of the rest
parameters - pros outweighs cons.