add result.nu module (#1176)
This is a module for storing and accessing previously output results.

The way it works is pretty simple: it uses the `display_output` hook to
append results to an environment variable. The subtlety that tripped me
up is that one must be very careful to preserve the metadata of `ls`
output to preserve the ls colors in the display output. The only
reliable way to do this is to store metadata separately, as most
operations will not preserve it.

I wasn't sure whether this would be better off in `modules` or `hooks`,
but it's structured as a module and provides the commands `result`,
`result ls` and `result select` for accessing the results, so I figured
it was better off in `modules`.