Extend dotnet completions documentation
.NET 10 introduces `dotnet completions script nushell`.

I took the opportunity to try and update the dotnet completions, to test if it works and what works, but ultimately concluded that it's incomplete/unusable as is.

I updated the README.md with my findings.

I assume `dotnet complete` is currently missing and will be extended with `dotnet completions` completions eventually.

I'm not sure whether the current extern export is preferable over an external completer configuration, but it seems plausibly preferable to me, because it doesn't mix with other external completer configurations and is used earlier in the completion chain before falling back to external completers. If that is the case, the dotnet command generates suboptimal completion configuration.