Remove direnv hook example (#1084)
As noted in https://github.com/nushell/nushell/issues/15452, having a
direnv example in nu_scripts and one in the Cookbook is a problem in
that they've rarely been in sync. Since there's *documentation* for this
in the Cookbook, let's remove it from nu_scripts and just have one going
forward.