panache-git: fix bugs showing modified files and non-home dirs (#1188)
This PR fixes two bugs in the panache-git prompt:

First, modified file counts were not being properly shown after updating
to Nushell v0.109, which contained a breaking change in `split column`:
https://www.nushell.sh/blog/2025-11-29-nushell_v0_109_0.html#switch-split-column-to-0-index-like-other-commands-toc

panache-git relied on the old default column names when splitting
columns to count modified files. This has been fixed with explicit
column names so that the default names no longer matter.

Second, when the current working directory was outside the home
directory, panache-git would throw an error.

The code was attempting to join path strings that didn't exist, and the
`str join` in question turned out to be unnecessary, so it was removed,
avoiding the error.